consistently rename lib.models.mirrors => lib.models.pacman
